+++ src/main/java/org/tinycloud/jdbc/page/SqlitePageHandleImpl.java
@@ -17,11 +17,11 @@ public class SqlitePageHandleImpl implements IPageHandle {
      * @return 处理过后的sql
      */
     @Override
-    public String handlerPagingSQL(String oldSQL, int pageNo, int pageSize) {
+    public String handlerPagingSQL(String oldSQL, long pageNo, long pageSize) {
         StringBuilder sql = new StringBuilder(oldSQL);
 
-        int offset = (pageNo - 1) * pageSize;
-        int limit = pageSize;
+        long offset = (pageNo - 1) * pageSize;
+        long limit = pageSize;
         if (limit > 0) {
             sql.append("\n LIMIT ").append(limit);
         }
